+++ zerver/lib/message.py
@@ -1012,7 +1012,9 @@ def get_starred_message_ids(user_profile: UserProfile) -> List[int]:
     )
 
 
-def get_raw_unread_data(user_profile: UserProfile) -> RawUnreadMessagesResult:
+def get_raw_unread_data(
+    user_profile: UserProfile, message_ids: Optional[List[int]] = None
+) -> RawUnreadMessagesResult:
     excluded_recipient_ids = get_inactive_recipient_ids(user_profile)
 
     user_msgs = (
@@ -1021,9 +1023,6 @@ def get_raw_unread_data(user_profile: UserProfile) -> RawUnreadMessagesResult:
         )
         .exclude(
             message__recipient_id__in=excluded_recipient_ids,
-        )
-        .extra(
-            where=[UserMessage.where_unread()],
         )
         .values(
             "message_id",
@@ -1036,6 +1035,16 @@ def get_raw_unread_data(user_profile: UserProfile) -> RawUnreadMessagesResult:
         )
         .order_by("-message_id")
     )
+
+    if message_ids is not None:
+        # When users are marking just a few messages as unread, we just need
+        # those ids, and we know they're unread.
+        user_msgs = user_msgs.filter(message_id__in=message_ids)
+    else:
+        # At page load we need all unread messages.
+        user_msgs = user_msgs.extra(
+            where=[UserMessage.where_unread()],
+        )
 
     # Limit unread messages for performance reasons.
     user_msgs = list(user_msgs[:MAX_UNREAD_MESSAGES])
@@ -1267,6 +1276,72 @@ def remove_message_id_from_unread_mgs(state: RawUnreadMessagesResult, message_id
     state["huddle_dict"].pop(message_id, None)
     state["unmuted_stream_msgs"].discard(message_id)
     state["mentions"].discard(message_id)
+
+
+def format_unread_message_details(
+    raw_unread_data: RawUnreadMessagesResult,
+) -> Dict[str, Dict[str, Any]]:
+    unread_data = {}
+
+    for message_id, private_message_details in raw_unread_data["pm_dict"].items():
+        message_details = {
+            "type": "private",
+            "sender_id": private_message_details["sender_id"],
+        }
+        if message_id in raw_unread_data["mentions"]:
+            message_details["mentioned"] = True
+        unread_data[str(message_id)] = message_details
+
+    for message_id, stream_message_details in raw_unread_data["stream_dict"].items():
+        if message_id in raw_unread_data["unmuted_stream_msgs"]:
+            unmuted_stream_msg = True
+        else:
+            unmuted_stream_msg = False
+
+        message_details = {
+            "type": "stream",
+            "stream_id": stream_message_details["stream_id"],
+            "topic": stream_message_details["topic"],
+            # Clients don't need this detail, but we need it internally for apply_events.
+            "unmuted_stream_msg": unmuted_stream_msg,
+        }
+        if message_id in raw_unread_data["mentions"]:
+            message_details["mentioned"] = True
+        unread_data[str(message_id)] = message_details
+
+    for message_id, huddle_message_details in raw_unread_data["huddle_dict"].items():
+        message_details = {
+            "type": "huddle",
+            "user_ids_string": huddle_message_details["user_ids_string"],
+        }
+        if message_id in raw_unread_data["mentions"]:
+            message_details["mentioned"] = True
+        unread_data[str(message_id)] = message_details
+
+    return unread_data
+
+
+def add_message_to_unread_msgs(
+    state: RawUnreadMessagesResult, message_id: int, message: Dict[Any, Any]
+) -> None:
+    if message.get("mentioned"):
+        state["mentions"].add(message_id)
+
+    if message["type"] == "private":
+        state["pm_dict"][message_id] = {
+            "sender_id": message["sender_id"],
+        }
+    elif message["type"] == "stream":
+        state["stream_dict"][message_id] = {
+            "stream_id": message["stream_id"],
+            "topic": message["topic"],
+        }
+        if message["unmuted_stream_msg"]:
+            state["unmuted_stream_msgs"].add(message_id)
+    elif message["type"] == "huddle":
+        state["huddle_dict"][message_id] = {
+            "user_ids_string": message["user_ids_string"],
+        }
 
 
 def estimate_recent_messages(realm: Realm, hours: int) -> int: